pyproject.toml
[tool.poetry]
name = "wonder"
version = "0.1.0"
description = ""
authors = ["Frieda (Jingying) Huang <jingyingfhuang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.39.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pyautogui = "^0.9.54"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
